+++ packages/server/stores/s3/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @walkeros/server-store-s3
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies [a6a0ea7]
+  - @walkeros/core@4.0.2
+
 ## 4.0.1
 
 ### Patch Changes